client-bank-plan/backend/requirements.txt
flask>=3.1
flask-sqlalchemy>=3.1
flask-cors>=5.0
flask-socketio>=5.5
dash>=3.0
dash-bootstrap-components>=2.0
plotly>=6.0
python-socketio>=5.12
eventlet>=0.37
